backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Vector DB and ML
pinecone-client==3.0.0
sentence-transformers==2.3.1
torch==2.1.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.2

# Web scraping
httpx==0.26.0
beautifulsoup4==4.12.3

# File monitoring
watchdog==4.0.0

# MCP Server
mcp==0.9.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
